Refactored to use a mixin, since we were replicating logic
